refactor: circuit module
